wrap images in figure tags, show caption on hover
This is just a proof of concept in case you like it. I noticed that
captions weren't showing up, and sort of went down the rabbit hole.
To see what it looks like, I modified backg_1_dowling.txt, so you
can run it and see what you think. As you'll see, the syntax for
adding these kinds of figures with hover captions is different from
normal markdown.