addresses #18 using footnotes for definitions
For these changes to take effect, the Markdown files will
have to be edited as shown in expl_1_mem.txt. This commit
can be reverted if the team would prefer a different implementation.
Downside of this one is that it inserts a footnote reference character.
Upside is that it requires minimal modification of the Markdown,
while still displaying all the definitions at the bottom of the text.